Remove the model.
