+++ datasource/jaxb/src/main/resources/inventory.xml
