+++ driver/main/pom.xml
@@ -1,0 +1,24 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <groupId>org.opennms.oce</groupId>
+        <artifactId>driver</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>org.opennms.oce.driver</groupId>
+    <artifactId>org.opennms.oce.driver.main</artifactId>
+    <name>OCE :: Driver :: Main</name>
+    <packaging>bundle</packaging>
+
+    <dependencies>
+         <dependency>
+             <groupId>org.opennms.oce.datasource</groupId>
+             <artifactId>org.opennms.oce.datasource.api</artifactId>
+         </dependency>
+        <dependency>
+            <groupId>org.opennms.oce.engine</groupId>
+            <artifactId>org.opennms.oce.engine.api</artifactId>
+        </dependency>
+    </dependencies>
+
+</project>
