+++ engine/topology/src/main/java/org/opennms/oce/engine/topology/model/OperationalState.java
@@ -1,0 +1,6 @@
+package org.opennms.oce.engine.topology.model;
+
+public enum OperationalState {
+    UNKNOWN, NORMAL, NSA, SA
+
+}
