+++ engine/topology/src/main/java/org/opennms/oce/engine/topology/model/ServiceState.java
@@ -1,0 +1,13 @@
+package org.opennms.oce.engine.topology.model;
+
+public enum ServiceState {
+    @Deprecated
+    IN,
+    @Deprecated
+    OUT,
+    UNDEFINED,
+    IN_MAINTENANCE,
+    COMMISSIONING,
+    IN_SERVICE,
+    OUT_OF_SERVICE,
+}
