+++ engine/topology/src/main/java/org/opennms/oce/engine/topology/shell/ModelTypes.java
@@ -1,0 +1,58 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2018 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.oce.engine.topology.shell;
+
+import java.util.Set;
+
+import org.apache.karaf.shell.api.action.Action;
+import org.apache.karaf.shell.api.action.Command;
+import org.apache.karaf.shell.api.action.lifecycle.Reference;
+import org.apache.karaf.shell.api.action.lifecycle.Service;
+import org.opennms.oce.model.api.Model;
+import org.opennms.oce.model.api.ModelBuilder;
+
+@Command(scope = "oce", name = "modelTypes", description="Model Types Listing")
+@Service
+public class ModelTypes implements Action {
+    @Reference
+    private ModelBuilder builder;
+
+    @Override
+    public Object execute() throws Exception {
+        Model model = builder.buildModel();
+        final Set<String> types = model.getTypes();
+        if (types == null || types.isEmpty()) {
+            System.out.println("No model types");
+        }
+        for (String type : types) {
+            System.out.println("ModelType: " + type);
+        }
+        return types;
+    }
+}
