--- model/api/pom.xml
@@ -1,12 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>org.opennms.oce</groupId>
-        <artifactId>model</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>org.opennms.oce.model</groupId>
-    <artifactId>org.opennms.oce.model.api</artifactId>
-    <name>OCE :: Model :: API</name>
-    <packaging>bundle</packaging>
-</project>
