--- model/api/src/main/java/org/opennms/oce/model/api/ModelObject.java
@@ -1,80 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.oce.model.api;
-
-import java.util.Set;
-
-import org.opennms.oce.model.alarm.api.Alarm;
-
-public interface ModelObject {
-
-    ModelObject getParent();
-
-    Set<ModelObject> getChildren();
-
-    Set<ModelObject> getPeers();
-
-    Set<ModelObject> getUncles(); // Dependents
-
-    Set<ModelObject> getNephews(); // Dependencies
-
-    Group getChildGroup(String objectType);
-
-    Group getPeerGroup(String objectType);
-
-    Group getNephewGroup(String objectType);
-
-    Group getUncleGroup(String objectType);
-
-    /**
-     * Get the groups that this ModelObject will propagate Alarms to.
-     */
-    Set<Group> getAlarmGroups();
-
-    String getType();
-
-    String getSubType();
-
-    String getId();
-
-    String getFriendlyName();
-
-    OperationalState getOperationalState();
-
-    void setOperationalState(OperationalState state);
-
-    ServiceState getServiceState();
-
-    void setServiceState(ServiceState state);
-
-    void onAlarm(Alarm alarm);
-
-    Set<Alarm> getAlarms();
-
-}
