--- model/api/src/main/java/org/opennms/oce/model/api/OperationalState.java
@@ -1,6 +1,0 @@
-package org.opennms.oce.model.api;
-
-public enum OperationalState {
-    UNKNOWN, NORMAL, NSA, SA
-
-}
