--- model/api/src/main/java/org/opennms/oce/model/api/ServiceState.java
@@ -1,13 +1,0 @@
-package org.opennms.oce.model.api;
-
-public enum ServiceState {
-    @Deprecated
-    IN,
-    @Deprecated
-    OUT,
-    UNDEFINED,
-    IN_MAINTENANCE,
-    COMMISSIONING,
-    IN_SERVICE,
-    OUT_OF_SERVICE,
-}
