--- model/impl/src/main/java/org/opennms/oce/model/impl/ModelObjectKey.java
@@ -1,68 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.oce.model.impl;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-
-public class ModelObjectKey {
-
-    private final List<String> tokens;
-
-    public static ModelObjectKey key(String... tokens) {
-        return new ModelObjectKey(tokens);
-    }
-
-    public ModelObjectKey(String... tokens) {
-        this.tokens = Arrays.asList(tokens);
-    }
-
-    public List<String> getTokens() {
-        return tokens;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        ModelObjectKey that = (ModelObjectKey) o;
-        return Objects.equals(tokens, that.tokens);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(tokens);
-    }
-
-    @Override
-    public String toString() {
-        return tokens.toString();
-    }
-}
