--- model/impl/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,26 +1,0 @@
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0"
-           xmlns:ext="http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.0.0">
-
-    <!-- Configuration properties -->
-    <cm:property-placeholder id="modelProperties" persistent-id="org.opennms.oce.model" update-strategy="reload">
-        <cm:default-properties>
-            <cm:property name="metamodelPath" value=""/>
-            <cm:property name="inventoryPath" value=""/>
-        </cm:default-properties>
-    </cm:property-placeholder>
-
-    <!-- Create and expose the builder -->
-    <bean id="modelBuilder" class="org.opennms.oce.model.impl.ModelBuilderImpl">
-        <argument ref="blueprintBundleContext"/>
-        <property name="metamodelPath" value="${metamodelPath}"/>
-        <property name="inventoryPath" value="${inventoryPath}"/>
-    </bean>
-    <service ref="modelBuilder" interface="org.opennms.oce.model.api.ModelBuilder"/>
-
-    <!-- Create and expose an instance of the model -->
-    <bean id="model" factory-ref="modelBuilder" factory-method="buildModel"/>
-    <service ref="model" interface="org.opennms.oce.model.api.Model"/>
-
-
-</blueprint>
