--- model/impl/src/test/resources/log4j2.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="WARN">
-  <Appenders>
-    <Console name="Console" target="SYSTEM_OUT">
-      <PatternLayout pattern="%d %-5p [%t] %c{1.}: %m%n"/>
-    </Console>
-  </Appenders>
-  <Loggers>
-    <Root level="DEBUG">
-      <AppenderRef ref="Console"/>
-    </Root>
-  </Loggers>
-</Configuration>
