--- model/pom.xml
@@ -1,19 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>org.opennms.oce</groupId>
-        <artifactId>oce-parent</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>model</artifactId>
-    <name>OCE :: Model</name>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>api</module>
-        <module>impl</module>
-        <module>shell</module>
-    </modules>
-</project>
-
