+++ pom.xml
@@ -31,9 +31,9 @@
 
     <modules>
         <module>datasource</module>
+        <module>driver</module>
         <module>engine</module>
         <module>features</module>
-        <module>model</module>
         <module>wrap</module>
     </modules>
